+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 # Contributing
 
-Contributions to `pydatasummary` are welcome and encouraged! The goal of this
+Contributions to `statsframe` are welcome and encouraged! The goal of this
 project is to make summarizing data and statistical models in python easier and
 more intuitive. If you have an idea for a new feature, or a bug fix, please make
 a suggestion. Every contribution is appreciated and will be considered.
