+++ pyproject.toml
@@ -1,13 +1,13 @@
 [tool.poetry]
-name = "pydatasummary"
-version = "0.0.2"
+name = "statsframe"
+version = "0.0.1"
 description = "Customizable data and model summaries in Python."
 authors = ["Niall Keleher <niall.keleher@gmail.com>"]
 packages = [{ include = "*", from = "src" }]
 license = "MIT"
 readme = "README.md"
-homepage = "https://github.com/NKeleher/pydatasummary#readme"
-repository = "https://github.com/NKeleher/pydatasummary"
+homepage = "https://github.com/NKeleher/statsframe#readme"
+repository = "https://github.com/NKeleher/statsframe"
 keywords = ["tables", "statistics", "econometrics"]
 classifiers = [
     # https://pypi.org/classifiers/
@@ -29,6 +29,7 @@ classifiers = [
 python = "^3.9"
 polars = "^0.20.5"
 pandas = "^2.1.4"
+importlib-metadata = "^7.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.4"
@@ -40,7 +41,6 @@ bandit = "^1.7.6"
 docformatter = "^1.7.5"
 mypy = "^1.8.0"
 jupyterlab = "^4.0.11"
-importlib-metadata = "^7.0.1"
 
 [build-system]
 requires = ["poetry-core"]
