+++ src/statsframe/__init__.py
@@ -1,12 +1,12 @@
-# Define pydatasummary version
+# Define statsframe version
 from importlib_metadata import version as _v
 
 # __version__ = "0.0.1"
-__version__ = _v("pydatasummary")
+__version__ = _v("statsframe")
 
 del _v
 
-# Import pydatasummary objects
+# Import statsframe objects
 # from ._tbl_data import *  # noqa: F401, F403, E402
 # from ._databackend import *  # noqa: F401, F403, E402
 from .ds import *  # noqa: F401, F403, E402
