+++ src/statsframe/_tbl_data.py
