+++ tests/test_skim.py
@@ -1,7 +1,7 @@
 import polars as pl
 from polars.testing import assert_frame_equal
 
-from pydatasummary.ds import skim
+from statsframe.ds import skim
 
 df = pl.DataFrame(
     {
